+++ tests/xharness/xharness.csproj
@@ -141,6 +141,7 @@
     <Compile Include="Jenkins\Reports\MarkdownReportWriter.cs" />
     <Compile Include="Jenkins\Reports\HtmlReportWriter.cs" />
     <Compile Include="Jenkins\RunSimulatorTasksFactory.cs" />
+    <Compile Include="Jenkins\TestServer.cs" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="..\..\tools\common\SdkVersions.cs">
